V 3.3.2.1-rc8 - es werden nun auch sehr lange MC-Nachrichten unterstützt Die neuen Oregon V3 Sensoren senden hintereinander 3 Nachrichten ohne Pause. Die empfangenen Pulse passen nicht in den message Puffer, nun werden auch die Manchesterdaten nach einem Überlauf verarbeitet.

+++ src/_micro-api/libraries/bitstore/src/bitstore.h
@@ -39,7 +39,7 @@ public:
 	void reset();
 	bool getByte(const uint8_t idx, uint8_t *retvalue);
 	uint8_t bytecount;  // Number of stored bytes
-	uint8_t valcount;  // Number of total values stored  --- todo uint16_t
+	uint16_t valcount;  // Number of total values stored
 	//uint8_t debug;
 	uint8_t bcnt;   // fuer debugzwecke nach public verschoben
